+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [4.5.3](https://github.com/dgiot/dgiot-dashboard/compare/v4.5.2...v4.5.3) (2022-05-09)
+## [4.5.3](https://github.com/dgiot/dgiot-dashboard/compare/v4.5.2...v4.5.3) (2022-05-10)
 
 
 ### Bug Fixes
@@ -51,6 +51,7 @@
 * **device:** rewrite device list interface ([0fdb3c8](https://github.com/dgiot/dgiot-dashboard/commit/0fdb3c8182ba074466d4fa5e33bf2bfab9014696))
 * **device:** rewrite device management interface ([85ad1b0](https://github.com/dgiot/dgiot-dashboard/commit/85ad1b072954fba850f9d0de27a280e6187e03de))
 * devicesDetailLite.vue ([78109fa](https://github.com/dgiot/dgiot-dashboard/commit/78109fa7bb9e248174167aff1509db6be94685a0))
+* **device:** task hidden ([a47151a](https://github.com/dgiot/dgiot-dashboard/commit/a47151a7c74839517699d106c4c7288a05ed56aa))
 * disable key ([fe3ebd6](https://github.com/dgiot/dgiot-dashboard/commit/fe3ebd665ef346eb66cb2c06a6e3d25b56609ede))
 * el-tabs ([a533430](https://github.com/dgiot/dgiot-dashboard/commit/a533430d77a5c9f4c8838b0844ac5850e982ae65))
 * factory details display page ([8d7cbc9](https://github.com/dgiot/dgiot-dashboard/commit/8d7cbc95a1a892c63827673c56fcf79d919a5460))